+++ Plugins/BetterImageViewer/CHANGELOG.md
@@ -1,4 +1,13 @@
 # [BetterImageViewer](https://1lighty.github.io/BetterDiscordStuff/?plugin=BetterImageViewer "BetterImageViewer") Changelog
+### 1.1.0
+- *SOME* people kept asking, so here you go.
+- Added image zoom, simply activate by click and holding on the image.
+- Use scroll wheel to zoom in and out, hold shift while scrolling to change lens size.
+- If you prefer using a different plugin to handle zooming for you, that's fine too, there is a toggle to disable it in settings.
+- The click and hold can of course be easily changed in the settings to your prefered method of zooming.
+![zoommode](https://i.imgur.com/A8HjQb9.png)
+- There are other options too, for those that may want it!
+![moreoptions](https://i.imgur.com/JGNe7Re.png)'
 ### 1.0.4
 - Fixed plugin refusing to work in PTB and canary